update all dependancies
