+++ packages/react-components/react-grid-layout/package.json
@@ -22,7 +22,7 @@
     "react-dom": "^16.12.0"
   },
   "devDependencies": {
-    "@types/react": "^16.9.19",
+    "@types/react": "^16.9.21",
     "@types/react-dom": "^16.9.5",
     "typescript": "^3.7.5"
   },
