+++ packages/react-hooks/use-is-in-viewport/package.json
@@ -23,7 +23,7 @@
   },
   "devDependencies": {
     "@types/debug": "^4.1.5",
-    "@types/react": "^16.9.19",
+    "@types/react": "^16.9.21",
     "debug": "^4.1.1",
     "typescript": "^3.7.5"
   },
