+++ packages/react-hooks/use-responsive-image-data/.gitignore
@@ -1,0 +1,3 @@
+.idea
+.cache
+node_modules
