+++ packages/react-hooks/use-responsive-image-data/.npmignore
@@ -1,0 +1,4 @@
+.idea
+.cache
+node_modules
+src
