+++ packages/react-hooks/use-responsive-image-data/README.md
@@ -1,0 +1,21 @@
+# @wbe/use-responsive-image-data
+
+This hook allow to return a specific IImage object depending of specific width value.
+
+![](https://img.shields.io/npm/v/@wbe/use-responsive-image-data/latest.svg)
+![](https://img.shields.io/bundlephobia/minzip/@wbe/use-responsive-image-data.svg)
+![](https://img.shields.io/david/willybrauner/libraries.svg?path=packages%2Freact-hooks%2Fuse-responsive-image-data)
+![](https://img.shields.io/npm/dt/@wbe/use-responsive-image-data.svg)
+![](https://img.shields.io/npm/l/@wbe/use-responsive-image-data.svg)
+
+## Installation
+
+```shell script
+$ npm install @wbe/use-responsive-image-data
+```
+
+## How to use
+
+```js
+import { useResponsiveImageData } from "@wbe/use-responsive-image-data";
+```
