+++ packages/react-hooks/use-responsive-image-data/tsconfig.json
