+++ packages/storybook/package.json
@@ -17,6 +17,7 @@
     "@wbe/react-responsive-image": "^0.0.1",
     "@wbe/use-bounding-client-rect": "^0.0.23",
     "@wbe/use-is-in-viewport": "^0.0.23",
+    "@wbe/use-responsive-image-data": "^0.0.1",
     "@wbe/use-window-size": "^0.0.23"
   },
   "devDependencies": {
@@ -30,9 +31,9 @@
     "@storybook/html": "^5.3.13",
     "@storybook/react": "^5.3.13",
     "@storybook/theming": "^5.3.13",
-    "@types/jest": "^25.1.2",
-    "@types/node": "^13.7.1",
-    "@types/react": "^16.9.19",
+    "@types/jest": "^25.1.3",
+    "@types/node": "^13.7.4",
+    "@types/react": "^16.9.21",
     "@types/react-dom": "^16.9.5",
     "awesome-typescript-loader": "^5.2.1",
     "babel-loader": "^8.0.6",
@@ -40,7 +41,7 @@
     "jest": "^25.1.0",
     "react-docgen-typescript-loader": "^3.6.0",
     "storybook-readme": "^5.0.8",
-    "ts-jest": "^25.2.0",
+    "ts-jest": "^25.2.1",
     "typescript": "^3.7.5"
   },
   "gitHead": "07a03e6afcdb266a9b142ec9f4202ef453387464"
